requirements.txt
# development requirements
pytest<6
requirements-parser<=0.2
